+++ src/main/java/UI/Tools.java
@@ -64,7 +64,7 @@ public class Tools {
     // Llenar tabla de cuentas bancarias
     public static void entablarCuentas(JTable tabla, int idCliente) {
         actualizarTabla(tabla, cuentaBancariaDAO.obtenerCuentasCliente(idCliente),
-                cuenta -> new Object[]{cuenta.getNumeroCuenta(), cuenta.getSaldo()},
+                cuenta -> new Object[]{(int) cuenta.getNumeroCuenta(), (double) cuenta.getSaldo()},
                 new Object[]{"No hay cuentas", "No hay cuentas"});
     }
